requirements.txt
anyio==3.6.2
bcrypt==4.0.1
bidict==0.22.0
cffi==1.15.1
click==8.1.3
cryptography==38.0.1
fastapi==0.85.1
Flask==2.2.2
Flask-Bcrypt==1.0.1
Flask-Cors==3.0.10
flask-marshmallow==0.14.0
Flask-SocketIO==5.3.1
Flask-SQLAlchemy==3.0.2
greenlet==1.1.3.post0
gunicorn==20.1.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
marshmallow==3.18.0
packaging==21.3
pycparser==2.21
pydantic==1.10.2
pyOpenSSL==22.1.0
pyparsing==3.0.9
PyJWT==1.7.1
python-dotenv==0.21.0
python-engineio==4.3.4
python-socketio==5.7.2
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.42
starlette==0.20.4
typing_extensions==4.4.0
Werkzeug==2.2.2
